etl/postgres_to_es/ETL/requirements.txt
psycopg2-binary==2.9.4
gunicorn==20.1.0
python-dotenv == 0.21.0
split_settings == 1.0.0
gunicorn==20.1.0
elasticsearch==7.7.0
elastic-transport==8.4.0
pydantic==1.10.2
